Moved AnnotationRoutingLoader to Service directory
