+++ src/Service/AnnotationRoutingLoader/AnnotationRoutingLoader.php
@@ -2,14 +2,14 @@
 
 declare(strict_types=1);
 
-namespace App\Router;
+namespace App\Service\AnnotationRoutingLoader;
 
 use ReflectionClass;
 use ReflectionMethod;
 use Symfony\Component\Routing\Loader\AnnotationClassLoader;
 use Symfony\Component\Routing\Route;
 
-class AnnotationLoader extends AnnotationClassLoader
+class AnnotationRoutingLoader extends AnnotationClassLoader
 {
     protected function configureRoute(Route $route, ReflectionClass $class, ReflectionMethod $method, object $annot): void
     {
